move cookie auth and ratelimit to theme
HiperToon also needs auth cookie
